+++ src/main/java/xueli/craftgame/renderer/model/TexturedModel.java
@@ -23,7 +23,7 @@ public class TexturedModel extends Module {
 		this.cubes.add(cube);
 	}
 
-	public int getRenderData(int x, int y, int z, byte face, Color color,float skyColor, FloatList buffer) {
+	public int getRenderData(float x, float y, float z, byte face, Color color,float skyColor, FloatList buffer) {
 		int vertCount = 0;
 		for (TexturedCube c : cubes) {
 			vertCount += c.getDrawData(x, y, z, face, color,skyColor, buffer);
